fix(ui): resolve a shell's session from every session, not the painted rail
Addresses the review on #328.

A search that matches shells but not agents leaves the rail's index without
a parent for those shells. Opening a second terminal from one then stored an
empty parent — and unlike a row, a stored parent does not come back when the
query clears: it is what the directory fallback reads later, on whichever
agent happens to share the directory. The spawn path resolves against every
session now; the rail keeps its filtered index, where a parent the view is
holding back should not nest a row.

The placement comment and the docs claimed a shell with no session sits in
its group, skipping the same-directory fallback that runs first — so a
terminal opened on a worktree's group row nests under the agent living
there, which the docs said it would not.

Tests move next to what they cover: the origin, name and index helpers to
terminal_test.go, the row's label rendering to listview_test.go, and the
placement assertions stay with the rail. Two comments that restated their
own code are gone; the one left on emit says why a shell has to sit
immediately after its session rather than what the loop does.
