+++ docs/usage.md
@@ -72,7 +72,7 @@ The shell is the `[tools.terminal]` block in [config.toml](configuration.md). It
 
 Shells gather under a **Terminals** rule pinned to the foot of the list, holding every shell from every group. The tree scrolls above it; the block keeps at most half the list and scrolls inside itself once you have more shells than that. A pinned row stands away from the tree, so it names the session it was opened for — the group would be shared by every worktree under it and tell them apart from nothing. A shell with no session to name falls back to the directory it was launched in.
 
-Settings (`s`) has a `terminal rows` row that switches this from `pinned` to `nested`, which hangs each shell off the session it was opened for, one level under it in the tree and marked with `❯` where an agent carries its status dot. The worktree a terminal belongs to is then the row above it rather than something to work out from a name, and the row leaves its own column empty because that row already says it. A second `T` on a shell joins it as a sibling instead of nesting deeper. A shell opened on a group row, which has no session to hang off, sits in that group beside the agents; so does one whose session has been archived or killed.
+Settings (`s`) has a `terminal rows` row that switches this from `pinned` to `nested`, which hangs each shell off the session it was opened for, one level under it in the tree and marked with `❯` where an agent carries its status dot. The worktree a terminal belongs to is then the row above it rather than something to work out from a name, and the row leaves its own column empty because that row already says it. A second `T` on a shell joins it as a sibling instead of nesting deeper. A shell that recorded no session — one opened on a group row, or one whose session has since been archived — looks for the oldest agent launched in its own directory, so a terminal opened on a worktree's group row still finds the agent living there. Only a shell that finds neither, in its own group, sits in that group beside the agents.
 
 A group's dots and counts describe its agents either way, so the shell you left running a build never shows up as work in progress.
